+++ src/EvalutionCS/Evalution.csproj
@@ -55,8 +55,10 @@
     <Compile Include="Ast\Expressions\MultiCalls\ObjectContextMethodCall.cs" />
     <Compile Include="Ast\Expressions\MultiCalls\ObjectContextPropertyCall.cs" />
     <Compile Include="Ast\UnaryOperator.cs" />
+    <Compile Include="ClassBuilder.generic.cs" />
     <Compile Include="ClassBuilder.cs" />
     <Compile Include="Context.cs" />
+    <Compile Include="EmitHelper.cs" />
     <Compile Include="Exceptions\EvalutionException.cs" />
     <Compile Include="Exceptions\InvalidNameException.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
